renamed steps page to allsteps, added empty slideshow page, updated buttons on overview page
