+++ client/src/Components/RecipeOverviewTab/index.tsx
@@ -36,18 +36,32 @@ const RecipeOverview: React.FC<Props> = ({ recipe }: Props) => {
             </div>
           ))}
         </div>
-        <Link to={'/recipe/' + recipe.slug + '/steps'}>
-          <button
-            className={
-              'w-24 mt-8 py-2 px-4 rounded-full ' +
-              colorScheme.buttonBg +
-              ' ' +
-              colorScheme.buttonText
-            }
-          >
-            To Recipe
-          </button>
-        </Link>
+        <div className='flex flex-row space-x-2'>
+          <Link to={'/recipe/' + recipe.slug + '/all-steps'}>
+            <button
+              className={
+                'w-48 mt-8 py-2 px-4 rounded-full ' +
+                colorScheme.buttonBg +
+                ' ' +
+                colorScheme.buttonText
+              }
+            >
+              View All Steps
+            </button>
+          </Link>
+          <Link to={'/recipe/' + recipe.slug + '/slideshow'}>
+            <button
+              className={
+                'w-48 mt-8 py-2 px-4 rounded-full ' +
+                colorScheme.buttonBg +
+                ' ' +
+                colorScheme.buttonText
+              }
+            >
+              View Slideshow
+            </button>
+          </Link>
+        </div>
       </div>
       <img className='w-80 rounded' src={recipe.finalShot} alt={recipe.name} />
       <div>
