+++ client/src/Pages/Recipe/slideshow.tsx
@@ -1,0 +1,5 @@
+import React from 'react'
+
+const RecipeSlideshow: React.FC = () => <div>Slideshow</div>
+
+export default RecipeSlideshow
